query language char - byte bug resolved
git-svn-id: svn://ciws.iasfbo.inaf.it/Ciws/DAS/trunk@46 2a20093d-90dd-4be6-b9fe-e465d393256e

+++ src/cpp/ql/bnfc/Parser.C
@@ -362,7 +362,7 @@ ListName* pListName(const char *str)
 #line 263 "grammar.y"
 typedef union
 {
-  int int_;
+  long long int_;
   char char_;
   double double_;
   char* string_;
